create token to facade
